fix: split client compute job subscription
